refact: terminator cli
